Use different strategy for API i18n
