+++ examples/tanstack-start/src/locales/en/messages.po
@@ -13,15 +13,30 @@ msgstr ""
 "Language-Team: \n"
 "Plural-Forms: \n"
 
+#. js-lingui-explicit-id
+#: src/routes/users.$userId.tsx:18
+msgid "Failed to fetch user"
+msgstr "Failed to fetch user"
+
+#. js-lingui-explicit-id
+#: src/routes/users.route.tsx:18
+msgid "Failed to fetch users"
+msgstr "Failed to fetch users"
+
+#. js-lingui-explicit-id
+#: src/routes/api/users.$id.ts:22
+msgid "User not found"
+msgstr "User not found"
+
 #: src/routes/posts_.$postId.deep.tsx:24
 msgid "← All Posts"
 msgstr "← All Posts"
 
-#: src/routes/deferred.tsx:57
+#: src/routes/deferred.tsx:70
 msgid "Count: {count}"
 msgstr "Count: {count}"
 
-#: src/routes/posts.$postId.tsx:31
+#: src/routes/posts.$postId.tsx:35
 msgid "Deep View"
 msgstr "Deep View"
 
@@ -29,15 +44,7 @@ msgstr "Deep View"
 msgid "Deferred"
 msgstr "Deferred"
 
-#: src/routes/users.$userId.tsx:16
-msgid "Failed to fetch user"
-msgstr "Failed to fetch user"
-
-#: src/routes/users.route.tsx:14
-msgid "Failed to fetch users"
-msgstr "Failed to fetch users"
-
-#: src/components/NotFound.tsx:15
+#: src/components/NotFound.tsx:19
 msgid "Go back"
 msgstr "Go back"
 
@@ -45,11 +52,11 @@ msgstr "Go back"
 msgid "Go Back"
 msgstr "Go Back"
 
-#: src/routes/_pathlessLayout/_nested-layout.tsx:19
+#: src/routes/_pathlessLayout/_nested-layout.tsx:21
 msgid "Go to route A"
 msgstr "Go to route A"
 
-#: src/routes/_pathlessLayout/_nested-layout.tsx:27
+#: src/routes/_pathlessLayout/_nested-layout.tsx:29
 msgid "Go to route B"
 msgstr "Go to route B"
 
@@ -58,31 +65,31 @@ msgstr "Go to route B"
 msgid "Home"
 msgstr "Home"
 
-#: src/routes/_pathlessLayout.tsx:11
+#: src/routes/_pathlessLayout.tsx:12
 msgid "I'm a layout"
 msgstr "I'm a layout"
 
-#: src/routes/_pathlessLayout/_nested-layout.tsx:11
+#: src/routes/_pathlessLayout/_nested-layout.tsx:12
 msgid "I'm a nested layout"
 msgstr "I'm a nested layout"
 
-#: src/routes/_pathlessLayout/_nested-layout/route-a.tsx:11
+#: src/routes/_pathlessLayout/_nested-layout/route-a.tsx:13
 msgid "I'm A!"
 msgstr "I'm A!"
 
-#: src/routes/_pathlessLayout/_nested-layout/route-b.tsx:11
+#: src/routes/_pathlessLayout/_nested-layout/route-b.tsx:13
 msgid "I'm B!"
 msgstr "I'm B!"
 
-#: src/routes/deferred.tsx:59
+#: src/routes/deferred.tsx:74
 msgid "Increment"
 msgstr "Increment"
 
-#: src/routes/deferred.tsx:41
+#: src/routes/deferred.tsx:44
 msgid "Loading person..."
 msgstr "Loading person..."
 
-#: src/routes/deferred.tsx:51
+#: src/routes/deferred.tsx:60
 msgid "Loading stuff..."
 msgstr "Loading stuff..."
 
@@ -90,7 +97,7 @@ msgstr "Loading stuff..."
 msgid "Non-existent Post"
 msgstr "Non-existent Post"
 
-#: src/routes/users.route.tsx:29
+#: src/routes/users.route.tsx:33
 msgid "Non-existent User"
 msgstr "Non-existent User"
 
@@ -98,7 +105,7 @@ msgstr "Non-existent User"
 msgid "Pathless Layout"
 msgstr "Pathless Layout"
 
-#: src/routes/posts.$postId.tsx:12
+#: src/routes/posts.$postId.tsx:14
 msgid "Post not found"
 msgstr "Post not found"
 
@@ -106,19 +113,19 @@ msgstr "Post not found"
 msgid "Posts"
 msgstr "Posts"
 
-#: src/routes/posts.index.tsx:9
+#: src/routes/posts.index.tsx:11
 msgid "Select a post."
 msgstr "Select a post."
 
-#: src/routes/users.index.tsx:9
+#: src/routes/users.index.tsx:11
 msgid "Select a user."
 msgstr "Select a user."
 
-#: src/components/NotFound.tsx:21
+#: src/components/NotFound.tsx:25
 msgid "Start Over"
 msgstr "Start Over"
 
-#: src/components/NotFound.tsx:8
+#: src/components/NotFound.tsx:10
 msgid "The page you are looking for does not exist."
 msgstr "The page you are looking for does not exist."
 
@@ -130,7 +137,7 @@ msgstr "This Route Does Not Exist"
 msgid "Try Again"
 msgstr "Try Again"
 
-#: src/routes/users.$userId.tsx:22
+#: src/routes/users.$userId.tsx:26
 msgid "User not found"
 msgstr "User not found"
 
@@ -138,6 +145,6 @@ msgstr "User not found"
 msgid "Users"
 msgstr "Users"
 
-#: src/routes/index.tsx:11
+#: src/routes/index.tsx:12
 msgid "Welcome Home!!!"
 msgstr "Welcome Home!!!"
