+++ examples/tanstack-start/src/routes/api/users.$id.ts
@@ -2,6 +2,7 @@ import { json } from "@tanstack/react-start"
 import { createAPIFileRoute } from "@tanstack/react-start/api"
 import axios from "redaxios"
 import type { User } from "../../utils/users"
+import { i18n } from "@lingui/core"
 
 export const APIRoute = createAPIFileRoute("/api/users/$id")({
   GET: async ({ request, params }) => {
@@ -18,7 +19,7 @@ export const APIRoute = createAPIFileRoute("/api/users/$id")({
       })
     } catch (e) {
       console.error(e)
-      return json({ error: "User not found" }, { status: 404 })
+      return json({ error: i18n._("User not found") }, { status: 404 })
     }
   },
 })
